+++ lib/src/codecs/g_year/year_encoder.dart
@@ -1,0 +1,11 @@
+import 'dart:convert';
+
+import '../../types/gregorian_year.dart';
+
+/// Converts a [GregorianYear] object to its XSD gYear string representation.
+class GregorianYearEncoder extends Converter<GregorianYear, String> {
+  const GregorianYearEncoder();
+
+  @override
+  String convert(GregorianYear input) => input.toString();
+}