+++ lib/src/types/types.dart
@@ -1,1 +1,2 @@
 export 'year_month.dart';
+export 'gregorian_year.dart';