+++ lib/xsd.dart
@@ -4,5 +4,3 @@
 library;
 
 export 'src/xsd_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.