+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - Implement `xsd:long` data type codec
 - Implement `xsd:language` data type codec
 - Implement `xsd:gYearMonth` codec and data type
+- Implement `xsd:gYear` codec and data type